better model guidance through probablity space constraints
itroduce a new field to the field extract type. This field is used to
feed the model extra information about the value that should be
extracted from the document context using an indirect approach. This is
done by forcing the model to ouput all allowed enum variants in case of
select fields or adding an extra format identiefier before value
extraction occurs.