change return type of `extract` function
extract can take any schema, but returns a FieldExtract which is not as
flexible as the grammar. Given that more complex queries may not be
mapped to FieldExtract. So changing the return type to Value allows more
flexible reusage of `extract` and puts the burdon on the caller to
convert value to the desired type